meminfo route
