+++ webapp/src/routes/metrics/meminfo/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+	import MemInfo from '$lib/components/metrics/MemInfo.svelte';
+</script>
+
+<MemInfo />
